Add nz back to the paper vae example.  Add a paper vae_inferred_shape version without nz
