+++ deepppl/tests/good/paper/vae_inferred_shape.stan
@@ -5,7 +5,6 @@ networks {
 }
 
 data {
-    int nz;
     int<lower=0, upper=1> x[28, 28];
 }
 parameters {
@@ -19,7 +18,7 @@ model {
 }
 
 guide {
-  real encoded[2, nz] = encoder(x);
+  real encoded[2, _] = encoder(x);
   real mu_z[_] = encoded[1];
   real sigma_z[_] = encoded[2];
   z ~ normal(mu_z, sigma_z);
